+++ mayan/apps/sources/serializers.py
@@ -2,22 +2,48 @@ from __future__ import unicode_literals
 
 import logging
 
+from django.utils.translation import ugettext_lazy as _
+
 from rest_framework import serializers
 from rest_framework.reverse import reverse
+
+from mayan.apps.documents.models.document_models import DocumentType
 
 from .models import StagingFolderSource, WebFormSource
 
 logger = logging.getLogger(__name__)
 
 
+class StagingFolderFileUploadSerializer(serializers.Serializer):
+    document_type = serializers.PrimaryKeyRelatedField(
+        label=_('Document type'), many=False,
+        queryset=DocumentType.objects.all(), read_only=False
+    )
+    expand = serializers.BooleanField(
+        default=False, label=_('Expand'), help_text=_(
+            'Upload a compressed file\'s contained files as individual '
+            'documents'
+        )
+    )
+
+
 class StagingFolderFileSerializer(serializers.Serializer):
     filename = serializers.CharField(max_length=255)
     image_url = serializers.SerializerMethodField()
     url = serializers.SerializerMethodField()
+    encoded_filename = serializers.CharField(max_length=255)
+    upload_url = serializers.SerializerMethodField()
 
     def get_image_url(self, obj):
         return reverse(
             viewname='rest_api:stagingfolderfile-image-view',
+            args=(obj.staging_folder.pk, obj.encoded_filename,),
+            request=self.context.get('request')
+        )
+
+    def get_upload_url(self, obj):
+        return reverse(
+            viewname='rest_api:stagingfolderfile-upload-view',
             args=(obj.staging_folder.pk, obj.encoded_filename,),
             request=self.context.get('request')
         )
@@ -34,7 +60,13 @@ class StagingFolderSerializer(serializers.HyperlinkedModelSerializer):
     files = serializers.SerializerMethodField()
 
     class Meta:
-        fields = ('files',)
+        extra_kwargs = {
+            'url': {'view_name': 'rest_api:stagingfolder-detail'},
+        }
+        fields = (
+            'delete_after_upload', 'enabled', 'files', 'folder_path', 'id',
+            'label', 'preview_height', 'preview_width', 'uncompress', 'url'
+        )
         model = StagingFolderSource
 
     def get_files(self, obj):